requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
atomicwrites==1.4.1
attrs==22.2.0
certifi==2025.1.31
charset-normalizer==2.0.12
colorama==0.4.5
idna==3.10
importlib-metadata==4.8.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.1.4
pytest==7.0.1
requests==2.27.1
selenium==3.141.0
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.20
zipp==3.6.0
